seperate dto object
