+++ hackle-android-sdk/src/main/java/io/hackle/android/internal/user/UserCohortFetcher.kt
@@ -54,24 +54,3 @@ internal class UserCohortFetcher(
     }
 }
 
-internal class UserCohortsRequestDto(
-    val identifiers: Map<String, String>
-)
-
-internal fun UserCohortsRequestDto.encodeBase64Url(): String {
-    return Base64.encodeToString(toJson().toByteArray(UTF_8), URL_SAFE or NO_WRAP)
-}
-
-internal class UserCohortsResponseDto(
-    val cohorts: List<UserCohortDto>
-)
-
-internal class UserCohortDto(
-    val identifier: IdentifierDto,
-    val cohorts: List<Long>
-)
-
-internal class IdentifierDto(
-    val type: String,
-    val value: String
-)
